change responsability of savage
